+++ python/cdp/openapi_client/models/evm_user_operation.py
@@ -23,6 +23,7 @@ from typing import Any, ClassVar, Dict, List, Optional
 from typing_extensions import Annotated
 from cdp.openapi_client.models.evm_call import EvmCall
 from cdp.openapi_client.models.evm_user_operation_network import EvmUserOperationNetwork
+from cdp.openapi_client.models.user_operation_receipt import UserOperationReceipt
 from typing import Optional, Set
 from typing_extensions import Self
 
@@ -35,7 +36,8 @@ class EvmUserOperation(BaseModel):
     calls: List[EvmCall] = Field(description="The list of calls in the user operation.")
     status: StrictStr = Field(description="The status of the user operation.")
     transaction_hash: Optional[Annotated[str, Field(strict=True)]] = Field(default=None, description="The hash of the transaction that included this particular user operation. This gets set after the user operation is broadcasted and the transaction is included in a block.", alias="transactionHash")
-    __properties: ClassVar[List[str]] = ["network", "userOpHash", "calls", "status", "transactionHash"]
+    receipts: Optional[List[UserOperationReceipt]] = Field(default=None, description="The list of receipts associated with the user operation.")
+    __properties: ClassVar[List[str]] = ["network", "userOpHash", "calls", "status", "transactionHash", "receipts"]
 
     @field_validator('user_op_hash')
     def user_op_hash_validate_regular_expression(cls, value):
@@ -107,6 +109,13 @@ class EvmUserOperation(BaseModel):
                 if _item_calls:
                     _items.append(_item_calls.to_dict())
             _dict['calls'] = _items
+        # override the default output from pydantic by calling `to_dict()` of each item in receipts (list)
+        _items = []
+        if self.receipts:
+            for _item_receipts in self.receipts:
+                if _item_receipts:
+                    _items.append(_item_receipts.to_dict())
+            _dict['receipts'] = _items
         return _dict
 
     @classmethod
@@ -123,7 +132,8 @@ class EvmUserOperation(BaseModel):
             "userOpHash": obj.get("userOpHash"),
             "calls": [EvmCall.from_dict(_item) for _item in obj["calls"]] if obj.get("calls") is not None else None,
             "status": obj.get("status"),
-            "transactionHash": obj.get("transactionHash")
+            "transactionHash": obj.get("transactionHash"),
+            "receipts": [UserOperationReceipt.from_dict(_item) for _item in obj["receipts"]] if obj.get("receipts") is not None else None
         })
         return _obj