+++ python/cdp/openapi_client/test/test_evm_smart_accounts_api.py
@@ -34,6 +34,13 @@ class TestEVMSmartAccountsApi(unittest.IsolatedAsyncioTestCase):
         """
         pass
 
+    async def test_create_spend_permission(self) -> None:
+        """Test case for create_spend_permission
+
+        Create a spend permission
+        """
+        pass
+
     async def test_get_evm_smart_account(self) -> None:
         """Test case for get_evm_smart_account
 
@@ -62,10 +69,24 @@ class TestEVMSmartAccountsApi(unittest.IsolatedAsyncioTestCase):
         """
         pass
 
+    async def test_list_spend_permissions(self) -> None:
+        """Test case for list_spend_permissions
+
+        List spend permissions
+        """
+        pass
+
     async def test_prepare_user_operation(self) -> None:
         """Test case for prepare_user_operation
 
         Prepare a user operation
+        """
+        pass
+
+    async def test_revoke_spend_permission(self) -> None:
+        """Test case for revoke_spend_permission
+
+        Revoke a spend permission
         """
         pass