+++ typescript/src/accounts/evm/toNetworkScopedEvmServerAccount.ts
@@ -16,6 +16,7 @@ import type {
   SendTransactionOptions,
   TransactionResult,
 } from "../../actions/evm/sendTransaction.js";
+import type { UseSpendPermissionOptions } from "../../actions/evm/spend-permissions/types.js";
 import type { AccountQuoteSwapOptions, AccountSwapOptions } from "../../actions/evm/swap/types.js";
 import type { TransferOptions } from "../../actions/evm/transfer/types.js";
 import type { WaitForFundOperationOptions } from "../../actions/waitForFundOperationReceipt.js";
@@ -23,6 +24,7 @@ import type {
   ListEvmTokenBalancesNetwork,
   SendEvmTransactionBodyNetwork,
 } from "../../openapi-client/index.js";
+import type { SpendPermissionNetworks } from "../../spend-permissions/types.js";
 import type { Address, TransactionRequestEIP1559 } from "../../types/misc.js";
 
 /**
@@ -189,6 +191,19 @@ export async function toNetworkScopedEvmServerAccount<Network extends string>(
     Object.assign(account, {
       swap: async (swapOptions: AccountSwapOptions) => {
         return options.account.swap(swapOptions);
+      },
+    });
+  }
+
+  if (isMethodSupportedOnNetwork("useSpendPermission", resolvedNetworkName)) {
+    Object.assign(account, {
+      __experimental_useSpendPermission: async (
+        spendPermissionOptions: Omit<UseSpendPermissionOptions, "network">,
+      ) => {
+        return options.account.__experimental_useSpendPermission({
+          ...spendPermissionOptions,
+          network: options.network as SpendPermissionNetworks,
+        });
       },
     });
   }