+++ typescript/src/actions/solana/types.ts
@@ -1,6 +1,7 @@
 import { TransferOptions } from "./transfer.js";
 import {
   RequestFaucetOptions,
+  SendTransactionOptions,
   SignatureResult,
   SignMessageOptions,
   SignTransactionOptions,
@@ -13,6 +14,9 @@ import {
 } from "../waitForFundOperationReceipt.js";
 import { SolanaFundOptions } from "./fund/fund.js";
 import { SolanaQuoteFundOptions } from "./fund/quoteFund.js";
+import { SendTransactionResult } from "./sendTransaction.js";
+
+import type { SignTransactionResult } from "./signTransaction.js";
 
 export type AccountActions = {
   /**
@@ -87,12 +91,50 @@ export type AccountActions = {
    * const transaction = Buffer.from(serializedTransaction).toString("base64");
    *
    * // When you want to sign a transaction, you can do so by address and base64 encoded transaction
-   * const { signature } = await account.signTransaction({
+   * const { signedTransaction } = await account.signTransaction({
    *   transaction,
    * });
    * ```
    */
-  signTransaction: (options: Omit<SignTransactionOptions, "address">) => Promise<SignatureResult>;
+  signTransaction: (
+    options: Omit<SignTransactionOptions, "address">,
+  ) => Promise<SignTransactionResult>;
+
+  /**
+   * Sends a transaction.
+   *
+   * @param {SendTransactionOptions} options - Parameters for sending the transaction.
+   * @param {string} options.address - The address to send the transaction for.
+   * @param {string} options.transaction - The transaction to send.
+   * @param {string} [options.idempotencyKey] - An idempotency key.
+   *
+   * @returns A promise that resolves to the transaction signature.
+   *
+   * @example
+   * ```ts
+   * // Create a Solana account
+   * const account = await cdp.solana.createAccount();
+   *
+   * // Add your transaction instructions here
+   * const transaction = new Transaction()
+   *
+   * // Make sure to set requireAllSignatures to false, since signing will be done through the API
+   * const serializedTransaction = transaction.serialize({
+   *   requireAllSignatures: false,
+   * });
+   *
+   * // Base64 encode the serialized transaction
+   * const transaction = Buffer.from(serializedTransaction).toString("base64");
+   *
+   * // When you want to sign a transaction, you can do so by address and base64 encoded transaction
+   * const { transactionSignature } = await account.sendTransaction({
+   *   transaction,
+   * });
+   * ```
+   */
+  sendTransaction: (
+    options: Omit<SendTransactionOptions, "address">,
+  ) => Promise<SendTransactionResult>;
 
   /**
    * Transfers SOL or SPL tokens between accounts